chore: rename and unify import/import-chain-db commands.
