+++ crates/amaru/src/bin/amaru/cmd/mod.rs
@@ -15,14 +15,14 @@
 use amaru_kernel::Point;
 
 pub(crate) mod daemon;
-pub(crate) mod import;
-pub(crate) mod import_chain_db;
+pub(crate) mod import_headers;
+pub(crate) mod import_ledger_state;
 
 /// Default path to the on-disk ledger storage.
 pub(crate) const DEFAULT_LEDGER_DB_DIR: &str = "./ledger.db";
 
 /// Default path to the on-disk chain storage.
-pub(crate) const DEFAULT_CHAIN_DATABASE_PATH: &str = "./chain.db";
+pub(crate) const DEFAULT_CHAIN_DB_DIR: &str = "./chain.db";
 
 /// Default path to pre-computed on-chain data needed for block header validation.
 pub(crate) const DEFAULT_DATA_DIR: &str = "./data";
